% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{occurrence_terms}
\alias{occurrence_terms}
\alias{event_terms}
\title{Select support functions}
\usage{
occurrence_terms()

event_terms()
}
\value{
A vector of terms that are acceptable (but not mandatory) for that
use case.
}
\description{
When creating a Darwin Core archive, it is often useful to select only those
fields that conform to the standard. These functions provide a vector of
terms that can be used in combination with \code{\link[dplyr:select]{dplyr::select()}} and
\code{\link[dplyr:reexports]{dplyr::any_of()}} to quickly select Darwin Core terms for the relevant
data type (events, occurrences, media).
}
\seealso{
\code{\link[=basisOfRecord_values]{basisOfRecord_values()}} or \code{\link[=countryCode_values]{countryCode_values()}} for valid entries
\emph{within} a field.
}
