% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_measurements.R
\name{add_measure_column}
\alias{add_measure_column}
\title{Add measurement column to dataframe in Darwin Core standard}
\usage{
add_measure_column(.df, column = NULL, type = NULL, unit = NULL)
}
\description{
tbd
}
