% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Dataset of supported Country Codes}
\format{
A tibble with 249 rows and 3 columns:
\describe{
\item{name}{ISO 3166-1 alpha-2 code, pointing to its ISO 3166-2 article.}
\item{code}{English short name officially used by the ISO 3166 Maintenance Agency (ISO 3166/MA).}
\item{year}{Year when alpha-2 code was first officially assigned.}
}
}
\source{
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{Wikipedia}.
}
\usage{
country_codes
}
\description{
This tibble gives the ISO 3166-1 alpha-2 codes for countries, which are
the standard for supplying the Darwin Core term \code{countryCode}.
}
\seealso{
\code{\link[=use_locality]{use_locality()}} for assigning \code{countryCode} within a tibble;
\code{\link[=countryCode_values]{countryCode_values()}} to return valid codes as a vector.
}
\keyword{datasets}
