% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwin_core_terms.R
\docType{data}
\name{darwin_core_terms}
\alias{darwin_core_terms}
\title{Dataset of supported Darwin Core terms}
\format{
A \code{tibble} containing valid Darwin Core Standard terms (206 rows x 6 columns).
Column descriptions are as follows:
\describe{
\item{class}{TDWG group that a term belongs to.}
\item{term}{Column header names that can be used in Darwin Core}
\item{url}{Stable url to information describing the term.}
\item{definition}{Human-readable definition of the term.}
\item{comments}{Further information from TDWG.}
\item{examples}{Examples of how the field should be populated.}
}
}
\source{
Slightly modified version of a
\href{https://github.com/tdwg/dwc/blob/master/vocabulary/term_versions.csv}{table supplied by TDWG}.
}
\usage{
darwin_core_terms
}
\description{
The Darwin Core Standard is maintained by Biodiversity Information Standards,
previously known as the Taxonomic Databases Working Group and known by the
acronym 'TDWG'. This \code{tibble} is the full list of supported terms,
current as of 2024-12-10.

Users can use \code{\link[=occurrence_terms]{occurrence_terms()}} and \code{\link[=event_terms]{event_terms()}} as convenience
functions to access these terms.
}
\seealso{
\code{\link[=occurrence_terms]{occurrence_terms()}} and \code{\link[=event_terms]{event_terms()}} to get terms for use in
\code{\link[dplyr:select]{dplyr::select()}}
}
\keyword{datasets}
