% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_coordinates_sf.R
\name{set_coordinates_sf}
\alias{set_coordinates_sf}
\title{Set, create or modify columns with \code{sf} spatial information using Darwin Core}
\usage{
set_coordinates_sf(.df, geometry = NULL, .keep = "unused")
}
\arguments{
\item{.df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{geometry}{the latitude/longitude coordinates in decimal degrees as \code{sf} \code{POINT} class}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{\link[dplyr:mutate]{dplyr::mutate()}}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}
}
\value{
A tibble with the requested fields added.
}
\description{
This function helps format standard location fields to a \code{tibble}. It differs
from \code{set_coordinates()} by accepting \code{sf} geometry columns of class \code{POINT}
as coordinates (rather than \code{numeric} lat/lon coordinates). The advantage
of using an \code{sf} geometry is that the Coordinate Reference System (CRS) is
automatically formatted into the required \code{geodeticDatum} column.
}
\seealso{
\code{\link[=set_coordinates]{set_coordinates()}} for providing numeric coordinates,
\code{\link[=set_locality]{set_locality()}} for providing text-based spatial information
}
