% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwin_core_terms.R
\docType{data}
\name{darwin_core_terms}
\alias{darwin_core_terms}
\title{Dataset of supported Darwin Core terms}
\format{
A tibble with 206 rows and 6 columns:
\describe{
\item{class}{TDWG group that a term belongs to.}
\item{term}{Column header names that can be used in Darwin Core}
\item{url}{Stable url to information describing the term.}
\item{definition}{Human-readable definition of the term.}
\item{comments}{Further information from TDWG.}
\item{examples}{Examples of how the field should be populated.}
}
}
\source{
Slightly modified version of a table supplied by TDWG at
\href{https://github.com/tdwg/dwc/blob/master/vocabulary/term_versions.csv}{this link}.
}
\usage{
darwin_core_terms
}
\description{
The Darwin Core Standard is maintained by Biodiversity Information Standards,
previously known as the Taxonomic Databases Working Group and known by the
acronym 'TDWG'. This tibble is the full list of supported terms,
current as at 2024-12-10.
}
\seealso{
\code{\link[=occurrence_fields]{occurrence_fields()}} and \code{\link[=event_fields]{event_fields()}} to get terms for use in
\verb{[dplyr::select()]}
}
\keyword{datasets}
