% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_locality.R
\name{set_locality}
\alias{set_locality}
\title{Set, create or modify columns with locality information using Darwin Core}
\usage{
set_locality(
  .df,
  continent = NULL,
  country = NULL,
  countryCode = NULL,
  stateProvince = NULL,
  locality = NULL,
  .keep = "unused"
)
}
\arguments{
\item{.df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{continent}{(string) Valid continent. See details.}

\item{country}{Valid country name. See \code{country_codes}.}

\item{countryCode}{Valid country code. See \code{country_codes}.}

\item{stateProvince}{A sub-national region.}

\item{locality}{A specific location, such as a property or address.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{\link[dplyr:mutate]{dplyr::mutate()}}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}
}
\value{
A tibble with the requested fields.
}
\description{
Locality information refers to a description of a place, rather than a
spatial coordinate. This function helps specify or modify columns
with locality information in a flexible way.

In practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for fields in
the Darwin Core Standard.
}
\details{
Example values are:
\itemize{
\item \code{continent} should be one of \code{"Africa"}, \code{"Antarctica"}, \code{"Asia"},
\code{"Europe"}, \code{"North America"}, \code{"Oceania"} or \code{"South America"}.
\item \code{countryCode} should be supplied according to the
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 ALPHA-2}
standard, \href{https://dwc.tdwg.org/list/#dwc_countryCode}{as per TDWG advice}.
}
}
\seealso{
\code{\link[=set_coordinates]{set_coordinates()}} for numeric spatial data
}
